# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NGX Voice Sales Agent API
  description: |
    API para el agente de ventas conversacional NGX. Este sistema utiliza IA avanzada
    para automatizar y optimizar el proceso de ventas de servicios NGX, incluyendo
    NGX AGENTS ACCESS y programas de Hybrid Coaching.
    
    ## Características principales:
    - Conversaciones de voz en tiempo real con IA
    - Detección inteligente de objeciones y necesidades
    - Optimización de flujo conversacional con ML
    - Análisis predictivo de conversión
    - Sistema de caché multi-capa para <20ms de latencia
    
    ## Autenticación:
    Todos los endpoints requieren autenticación JWT Bearer token, excepto `/auth/login` y `/auth/register`.
    
  version: 1.0.0
  contact:
    name: NGX Development Team
    email: dev@ngx.com
  license:
    name: Proprietary
servers:
  - url: https://api.ngx-sales.com/v1
    description: Production server
  - url: https://staging-api.ngx-sales.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Iniciar sesión
      description: Autentica un usuario y retorna tokens JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin@ngx.com
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
      responses:
        200:
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        429:
          $ref: '#/components/responses/RateLimitError'
  /auth/register:
    post:
      tags:
        - auth
      summary: Registrar nuevo usuario
      description: Crea una nueva cuenta de usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        201:
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        409:
          description: El usuario ya existe
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Refrescar token
      description: Obtiene un nuevo access token usando el refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        200:
          description: Token refrescado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags:
        - auth
      summary: Obtener perfil del usuario actual
      description: Retorna la información del usuario autenticado
      responses:
        200:
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        401:
          $ref: '#/components/responses/UnauthorizedError'

  # Conversation endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    UnauthorizedError:
      description: Token de autenticación faltante o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequestError:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimitError:
      description: Límite de tasa excedido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Segundos hasta expiración
          example: 3600
    UserRegistration:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
        full_name:
          type: string
          example: Juan Pérez
        company:
          type: string
          example: NGX Fitness Studio
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        company:
          type: string
        role:
          type: string
          enum: [admin, agent, viewer]
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time

    # Conversation schemas
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Invalid authentication credentials
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time